Unify camera and doorbell implementation
