+++ drivers/doorbell/driver.compose.json
@@ -4,5 +4,13 @@
   "name": {
     "en": "Doorbell",
     "nl": "Deurbel"
+  },
+  "capabilities": ["alarm_motion", "alarm_sound"],
+  "capabilitiesOptions": {
+    "alarm_motion": {
+      "title": {
+        "en": "Motion Detected"
+      }
+    }
   }
 }
