+++ lib/camera/TuyaCameraConstants.ts
