+++ lib/camera/translations.json
